% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_classification.R
\name{get_classifications}
\alias{get_classifications}
\title{Get all Orphanet classifications}
\usage{
get_classifications(
  lang = c("CS", "DE", "EN", "ES", "FR", "IT", "NL", "PL", "PT"),
  api_key,
  output = c("df", "list", "json"),
  verbosity = 0
)
}
\arguments{
\item{lang}{The language in which entities are to be returned. Possible
values are "CS", "DE", "EN", "ES", "FR", "IT", "NL", "PL", "PT".}

\item{api_key}{The API key to use for authentication.}

\item{output}{The format of the output. Must be "df" for a data frame,
"list" for a list or "json" for a raw json string.}

\item{verbosity}{How much information to print}
}
\value{
A data frame containing all classifications in the chosen language.
}
\description{
Retrieves all clinical classifications from the 'ORPHAcode' API.
}
\examples{
\dontshow{if (identical(Sys.getenv("IN_PKGDOWN"), "true")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
get_classifications(lang = "EN",
                   api_key = "jfeldhege")
\dontshow{\}) # examplesIf}
}
