% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_entities.R
\name{get_entities}
\alias{get_entities}
\title{Title}
\usage{
get_entities(
  lang = c("CS", "DE", "EN", "ES", "FR", "IT", "NL", "PL", "PT"),
  apiKey,
  output = c("df", "list", "json"),
  verbosity = 0
)
}
\arguments{
\item{lang}{The language in which entities are to be returned. Possible
values are "CS", "DE", "EN", "ES", "FR", "IT", "NL", "PL", "PT".}

\item{apiKey}{The API key to use for authentication.}

\item{output}{The format of the output. Must be "df" for a data frame,
"list" for a list or "json" for a raw json string.}

\item{verbosity}{How much information to print}
}
\value{
A data frame containing all entities in the chosen language.
}
\description{
Title
}
\examples{
\dontrun{
get_entities(lang = "EN",
             apiKey = "Apikey here")
}
}
