% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_icd.R
\name{get_icd10}
\alias{get_icd10}
\title{Find association of ORPHAcode and ICD-10 code}
\usage{
get_icd10(
  ORPHAcode = NULL,
  icd10code = NULL,
  lang = c("CS", "DE", "EN", "ES", "FR", "IT", "NL", "PL", "PT"),
  apiKey,
  output = c("df", "list", "json"),
  verbosity = 0
)
}
\arguments{
\item{ORPHAcode}{The ORPHAcode}

\item{icd10code}{The ICD-10 code to look for}

\item{lang}{The language in which entities are to be returned. Possible
values are "CS", "DE", "EN", "ES", "FR", "IT", "NL", "PL", "PT".}

\item{apiKey}{The API key to use for authentication.}

\item{output}{The format of the output. Must be "df" for a data frame,
"list" for a list or "json" for a raw json string.}

\item{verbosity}{How much information to print}
}
\value{
A data frame
}
\description{
Find association of ORPHAcode and ICD-10 code
}
\examples{
\dontrun{
get_icd10(ORPHAcode = 15,
lang = "EN",
apiKey = "Apikey here")

get_icd10(icd10code = "Q77.4",
lang = "EN",
apiKey = "Apikey here")
}
}
\seealso{
\code{\link[=get_icd11]{get_icd11()}}
}
