% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_icd.R
\name{get_icd11}
\alias{get_icd11}
\title{Find association of ORPHAcode and ICD-11 code}
\usage{
get_icd11(
  ORPHAcode = NULL,
  icd11code = NULL,
  lang = c("CS", "DE", "EN", "ES", "FR", "IT", "NL", "PL", "PT"),
  apiKey,
  output = c("df", "list", "json"),
  verbosity = 0
)
}
\arguments{
\item{ORPHAcode}{The ORPHAcode}

\item{icd11code}{The ICD-11 code to look for}

\item{lang}{The language in which entities are to be returned. Possible
values are "CS", "DE", "EN", "ES", "FR", "IT", "NL", "PL", "PT".}

\item{apiKey}{The API key to use for authentication.}

\item{output}{The format of the output. Must be "df" for a data frame,
"list" for a list or "json" for a raw json string.}

\item{verbosity}{How much information to print}
}
\value{
Depending on the output format chosen, a data frame, list or raw json string.
}
\description{
Find association of ORPHAcode and ICD-11 code
}
\examples{
\dontrun{
get_icd11(ORPHAcode = 15,
lang = "EN",
apiKey = "Apikey here")

get_icd11(icd11code = "LD24.00",
lang = "EN",
apiKey = "Apikey here")
}
}
