% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{get_info}
\alias{get_info}
\title{Get information for an Orphacode}
\usage{
get_info(
  code,
  type = c("Definition", "Status", "Typology", "Classification", "ClassificationLevel"),
  lang = c("CS", "DE", "EN", "ES", "FR", "IT", "NL", "PL", "PT"),
  api_key,
  output = c("df", "list", "json"),
  verbosity = 0
)
}
\arguments{
\item{code}{Orphacode to search for. Must be numeric}

\item{type}{The type of information to return. Possible values are "Status",
"Definition", "Typology"}

\item{lang}{The language in which entities are to be returned. Possible
values are "CS", "DE", "EN", "ES", "FR", "IT", "NL", "PL", "PT".}

\item{api_key}{The API key to use for authentication.}

\item{output}{The format of the output. Must be "df" for a data frame,
"list" for a list or "json" for a raw json string.}

\item{verbosity}{How much information to print}
}
\value{
A data frame containing the ORPHAcode, the type and the date.
}
\description{
Information can be a definition, status or typology for an Orphacode
}
\examples{
\dontshow{if (identical(Sys.getenv("IN_PKGDOWN"), "true")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Get status for a code
get_info(code = 16,
         type = "Status",
         lang = "EN",
         api_key = "jfeldhege/oRphacode")

# Get definition for a code
get_info(code = 16,
         type = "Definition",
         lang = "EN",
         api_key = "jfeldhege/oRphacode")

\dontshow{\}) # examplesIf}
}
