% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{get_status}
\alias{get_status}
\title{Get status for a specified Orphacode}
\usage{
get_status(
  code,
  lang = c("CS", "DE", "EN", "ES", "FR", "IT", "NL", "PL", "PT"),
  api_key,
  output = c("df", "list", "json"),
  verbosity = 0
)
}
\arguments{
\item{code}{Orphacode to search for. Must be numeric}

\item{lang}{The language in which entities are to be returned. Possible
values are "CS", "DE", "EN", "ES", "FR", "IT", "NL", "PL", "PT".}

\item{api_key}{The API key to use for authentication.}

\item{output}{The format of the output. Must be "df" for a data frame,
"list" for a list or "json" for a raw json string.}

\item{verbosity}{How much information to print}
}
\value{
A data frame, list or json containing the status and date for a specified ORPHAcode.
}
\description{
Returns the status (active/inactive)
}
\examples{
\dontshow{if (identical(Sys.getenv("IN_PKGDOWN"), "true")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Get status for blue cone monochromatism (ORPHAcode 16)
get_status(code = 16,
         lang = "EN",
         api_key = "jfeldhege/oRphacode")
\dontshow{\}) # examplesIf}
}
