% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_funs.R
\name{search_code}
\alias{search_code}
\title{Search by Orphacode}
\usage{
search_code(
  code,
  lang = c("CS", "DE", "EN", "ES", "FR", "IT", "NL", "PL", "PT"),
  apiKey,
  output = c("df", "list", "json"),
  verbosity = 0
)
}
\arguments{
\item{code}{Orphacode to search for. Must be numeric}

\item{lang}{The language in which entities are to be returned. Possible
values are "CS", "DE", "EN", "ES", "FR", "IT", "NL", "PL", "PT".}

\item{apiKey}{The API key to use for authentication.}

\item{output}{The format of the output. Must be "df" for a data frame,
"list" for a list or "json" for a raw json string.}

\item{verbosity}{How much information to print}
}
\value{
A data frame containing
}
\description{
Search by Orphacode
}
\examples{
\dontrun{
search_code(code = 16,
            lang = "EN",
            apiKey = "Apikey here")
}

}
