% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_funs.R
\name{search_name}
\alias{search_name}
\title{Search by name}
\usage{
search_name(
  name,
  exact = FALSE,
  lang = c("CS", "DE", "EN", "ES", "FR", "IT", "NL", "PL", "PT"),
  api_key,
  output = c("df", "list", "json"),
  verbosity = 0
)
}
\arguments{
\item{name}{Character. The name to search for.}

\item{exact}{Logical. If true, only exact matches are returned. Otherwise an approximate, fuzzy search is made.}

\item{lang}{The language in which entities are to be returned. Possible
values are "CS", "DE", "EN", "ES", "FR", "IT", "NL", "PL", "PT".}

\item{api_key}{The API key to use for authentication.}

\item{output}{The format of the output. Must be "df" for a data frame,
"list" for a list or "json" for a raw json string.}

\item{verbosity}{How much information to print}
}
\value{
A data frame containing all matches for the searched name. Columns are:
\itemize{
\item Preferred term
\item ORPHAcode
\item Date
}
}
\description{
Search by name
}
\examples{
\dontshow{if (identical(Sys.getenv("IN_PKGDOWN"), "true")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Search for the name Marfan syndrome
search_name(name = "Marfan syndrome",
            exact = FALSE,
            lang = "EN",
            api_key = "jfeldhege/oRphacode")
\dontshow{\}) # examplesIf}
}
