% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_funs.R
\name{search_synonym}
\alias{search_synonym}
\title{Search by synonym}
\usage{
search_synonym(
  synonym,
  exact = FALSE,
  lang = c("CS", "DE", "EN", "ES", "FR", "IT", "NL", "PL", "PT"),
  apiKey,
  output = c("df", "list", "json"),
  verbosity = 0
)
}
\arguments{
\item{synonym}{The synonym to search for}

\item{exact}{Logical. If true, only exact matches are returned. Otherwise an approximate, fuzzy search is made.}

\item{lang}{The language in which entities are to be returned. Possible
values are "CS", "DE", "EN", "ES", "FR", "IT", "NL", "PL", "PT".}

\item{apiKey}{The API key to use for authentication.}

\item{output}{The format of the output. Must be "df" for a data frame,
"list" for a list or "json" for a raw json string.}

\item{verbosity}{How much information to print}
}
\value{
A data frame containing all matches for the searched synonym. Columns are:
\itemize{
\item Preferred term
\item ORPHAcode
\item Date
}
}
\description{
Search by synonym
}
\examples{
\dontrun{
search_synonym(synonym = "MFS",
               exact = FALSE,
               lang = "EN",
               apiKey = "Apikey here")
}

}
