% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_icd.R
\name{get_icd11}
\alias{get_icd11}
\title{Find association of ORPHAcode and ICD-11 code}
\usage{
get_icd11(
  orpha_code = NULL,
  icd11_code = NULL,
  lang = c("CS", "DE", "EN", "ES", "FR", "IT", "NL", "PL", "PT"),
  api_key,
  output = c("df", "list", "json"),
  verbosity = 0
)
}
\arguments{
\item{orpha_code}{The ORPHAcode}

\item{icd11_code}{The ICD-11 code to look for}

\item{lang}{The language in which entities are to be returned. Possible
values are "CS", "DE", "EN", "ES", "FR", "IT", "NL", "PL", "PT".}

\item{api_key}{The API key to use for authentication.}

\item{output}{The format of the output. Must be "df" for a data frame,
"list" for a list or "json" for a raw json string.}

\item{verbosity}{How much information to print}
}
\value{
Depending on the output format chosen, a data frame, list or raw json string.
}
\description{
Find association of ORPHAcode and ICD-11 code
}
\examples{
\dontshow{if (identical(Sys.getenv("IN_PKGDOWN"), "true")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

# Get ICD-11 code for Achondroplasia (ORPHAcode = 15)
get_icd11(orpha_code = 15,
          lang = "EN",
          api_key = "jfeldhege/oRphacode")

# Search orphacodes for ICD-11 code LD24.00 for Achondroplasia
get_icd11(icd11code = "LD24.00",
          lang = "EN",
          api_key = "jfeldhege/oRphacode")
\dontshow{\}) # examplesIf}
}
